[ovsp4rt_v3] Extract tx_acc_vsi_table_entry utilities
